Extract accessibility element collection into reusable collector
